fix: rainbowkit demo
