+++ examples/with-evm-rainbowkit/package.json
@@ -13,8 +13,7 @@
     "react-scripts": "5.0.1",
     "wagmi": "1.3.10",
     "viem": "1.5.4",
-    "web-vitals": "^2.1.4",
-    "web3": "^4.0.3"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
